+++ package.json
@@ -4,7 +4,7 @@
   "description": "Turns CSVs into graphs in a few simple steps; embeds onto other websites!",
   "main": "src/PublicLab.Grapher.js",
   "scripts": {
-    "build": "watch 'browserify src/PublicLab.Grapher.js -o dist/PublicLab.Grapher.js' src",
+    "build": "watch 'browserify src/SimpleDataGrapher.js -o dist/PublicLab.Grapher.js' src",
     "test": "mocha spec/js/"
   },
   "repository": {
